+++ CHANGELOG.md
@@ -6,6 +6,14 @@ Gotta blame the broken mods!
 ------------------------------------------------
        | Blame Fabric changelog |
        
+       
+   (V.2.2.2 Changes) (1.16.2/1.16.3/1.16.4 Minecraft)
+      
+    Structure Spacing:
+•  Now will detect if separation value is less than or equal 
+   to spacing value because that will crash structures.
+   
+   
    (V.2.2.1 Changes) (1.16.2/1.16.3/1.16.4 Minecraft)
       
     Missing Loot Table: